requirements.txt
PyAudioWPatch>=0.2.12
google-cloud-speech>=2.20.0
openai>=1.30.0
PyMuPDF>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
Pillow>=10.0.0
mss>=9.0.0
imagehash>=4.3.0
rapidfuzz>=3.0.0
pyinstaller>=6.0.0
